requirements.txt
groq
pinecone
sentence-transformers
streamlit
python-dotenv
requests
